move component to a more fitting location
