--- src/app/pages/terminate-test-modal/terminate-test-modal.routing.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { TerminateTestModal } from '@pages/terminate-test-modal/terminate-test-modal';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: TerminateTestModal,
-  },
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class TerminateTestModalRoutingModule {}
