+++ src/components/common/common-components.module.ts
@@ -23,6 +23,7 @@ import { SignatureComponent } from '@components/common/signature/signature';
 import { TransmissionDisplayComponent } from '@components/common/transmission-display/transmission-display';
 import { ActivityCodeComponent } from '@components/common/activity-code/activity-code';
 import { CPCDebriefCardComponent } from '@components/common/cpc-debrief-card/cpc-debrief-card';
+import { TerminateTestModalModule } from '@components/common/terminate-test-modal/terminate-test-modal.module';
 import { TypeaheadDropdownComponent } from './typeahead-dropdown/typeahead-dropdown';
 import { WarningBannerComponent } from './warning-banner/warning-banner';
 import { TabsComponent } from './tabs/tabs';
@@ -82,6 +83,7 @@ import { VRNCaptureModalModule } from './vrn-capture-modal/vrn-capture-modal.mod
     ReactiveFormsModule,
     VRNCaptureModalModule,
     PipesModule,
+    TerminateTestModalModule,
   ],
   exports: [
     SignatureAreaComponent,
