+++ src/components/common/terminate-test-modal/terminate-test-modal.scss
@@ -1,4 +1,4 @@
-@import "src/theme/sass-partials/modal";
+@import "../../../theme/sass-partials/modal";
 
 #terminate-test-modal-card {
   padding: 0 0 0 0;
